virDomainJobObj: Use 'unsigned int' instead of 'unsigned long' for 'apiFlags' field
The callers store only an 'unsigned int' in the field. Convert it to the
proper type including parser/formatter.

+++ src/qemu/qemu_migration_params.h
@@ -98,7 +98,7 @@ int
 qemuMigrationParamsApply(virDomainObj *vm,
                          int asyncJob,
                          qemuMigrationParams *migParams,
-                         unsigned long apiFlags);
+                         unsigned int apiFlags);
 
 int
 qemuMigrationParamsEnableTLS(virQEMUDriver *driver,
@@ -145,7 +145,7 @@ void
 qemuMigrationParamsReset(virDomainObj *vm,
                          int asyncJob,
                          qemuMigrationParams *origParams,
-                         unsigned long apiFlags);
+                         unsigned int apiFlags);
 
 void
 qemuMigrationParamsFormat(virBuffer *buf,
